backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pyodbc==5.0.1
python-dotenv==1.0.0
dropbox==11.36.2
schedule==1.2.1
pydantic==2.5.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
aiosmtplib==2.0.2
